# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoltIQ",
    "description": "VoltIQ is an enterprise-grade AI-powered fleet electrification and asset intelligence platform.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/fleet/summary": {
      "get": {
        "tags": [
          "Fleet Electrification Readiness"
        ],
        "summary": "Get Fleet Readiness Summary",
        "description": "Get a summary of the electrification readiness status of the entire fleet,\ndynamically calculated from the fleet_readiness dataset, along with sample recommendations.",
        "operationId": "get_fleet_readiness_summary_api_fleet_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetSummaryOutput"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/fleet/vehicle/{vehicle_id}": {
      "get": {
        "tags": [
          "Fleet Electrification Readiness"
        ],
        "summary": "Get Vehicle Readiness",
        "description": "Get detailed EV replacement feasibility for a specific fleet vehicle.",
        "operationId": "get_vehicle_readiness_api_fleet_vehicle__vehicle_id__get",
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetVehicleReadiness"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fleet/predict": {
      "post": {
        "tags": [
          "Fleet Electrification Readiness"
        ],
        "summary": "Predict Fleet Readiness",
        "description": "Predict EV readiness score and transition category for any vehicle given its\nmechanical and operational telemetry.",
        "operationId": "predict_fleet_readiness_api_fleet_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FleetPredictInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetPredictOutput"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/fleet/summary": {
      "get": {
        "tags": [
          "Fleet Electrification Readiness"
        ],
        "summary": "Get Fleet Readiness Summary",
        "description": "Get a summary of the electrification readiness status of the entire fleet,\ndynamically calculated from the fleet_readiness dataset, along with sample recommendations.",
        "operationId": "get_fleet_readiness_summary_api_v1_fleet_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetSummaryOutput"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fleet/vehicle/{vehicle_id}": {
      "get": {
        "tags": [
          "Fleet Electrification Readiness"
        ],
        "summary": "Get Vehicle Readiness",
        "description": "Get detailed EV replacement feasibility for a specific fleet vehicle.",
        "operationId": "get_vehicle_readiness_api_v1_fleet_vehicle__vehicle_id__get",
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetVehicleReadiness"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fleet/predict": {
      "post": {
        "tags": [
          "Fleet Electrification Readiness"
        ],
        "summary": "Predict Fleet Readiness",
        "description": "Predict EV readiness score and transition category for any vehicle given its\nmechanical and operational telemetry.",
        "operationId": "predict_fleet_readiness_api_v1_fleet_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FleetPredictInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetPredictOutput"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/models/registry": {
      "get": {
        "tags": [
          "Model Registry"
        ],
        "summary": "Get Model Registry",
        "description": "Fetch the list of all registered models, their metadata, feature schemas,\nand latest performance evaluation metrics.",
        "operationId": "get_model_registry_api_v1_models_registry_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRegistryOutput"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{model_id}": {
      "get": {
        "tags": [
          "Model Registry"
        ],
        "summary": "Get Model Details",
        "description": "Fetch extensive metadata, training parameters, and metrics for a specific model ID.",
        "operationId": "get_model_details_api_v1_models__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfoOutput"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Detailed description of the error"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Unique UUID associated with the request log"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "example": {
          "detail": "Requested resource was not found.",
          "request_id": "8f8c7b8c-5544-4321-9abc-def012345678"
        }
      },
      "FleetPredictInput": {
        "properties": {
          "vehicle_age_years": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Vehicle Age Years",
            "description": "Age of vehicle in years"
          },
          "usage_hours": {
            "type": "number",
            "minimum": 0.0,
            "title": "Usage Hours",
            "description": "Total engine/usage hours"
          },
          "fuel_consumption": {
            "type": "number",
            "minimum": 0.0,
            "title": "Fuel Consumption",
            "description": "Fuel consumption rate (L/100km)"
          },
          "health_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Health Score",
            "description": "Vehicle health index"
          },
          "load_capacity": {
            "type": "number",
            "minimum": 0.0,
            "title": "Load Capacity",
            "description": "Max load capacity in kg",
            "default": 5000.0
          },
          "actual_load": {
            "type": "number",
            "minimum": 0.0,
            "title": "Actual Load",
            "description": "Typical load carried in kg",
            "default": 2500.0
          },
          "load_utilization_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Load Utilization Pct",
            "description": "Load capacity utilisation %",
            "default": 50.0
          },
          "fuel_per_hour": {
            "type": "number",
            "minimum": 0.0,
            "title": "Fuel Per Hour",
            "description": "Fuel burned per hour",
            "default": 1.5
          },
          "maintenance_cost": {
            "type": "number",
            "minimum": 0.0,
            "title": "Maintenance Cost",
            "description": "Annual maintenance cost in USD",
            "default": 1000.0
          },
          "days_since_last_maintenance": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Days Since Last Maintenance",
            "description": "Days since last service",
            "default": 30
          },
          "failure_history": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Failure History",
            "description": "Count of historical mechanical failures",
            "default": 0
          },
          "anomalies_detected": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Anomalies Detected",
            "description": "Sensor anomalous flag counts",
            "default": 0
          },
          "diagnostic_trouble_code_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Diagnostic Trouble Code Count",
            "description": "Active trouble codes count",
            "default": 0
          },
          "predictive_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Predictive Score",
            "description": "Maintenance predictive score",
            "default": 0.5
          },
          "pcr": {
            "type": "number",
            "title": "Pcr",
            "description": "Performance Cost Ratio",
            "default": 0.5
          },
          "uir": {
            "type": "number",
            "title": "Uir",
            "description": "Utilisation Idle Ratio",
            "default": 0.5
          },
          "tpi": {
            "type": "number",
            "title": "Tpi",
            "description": "Total Performance Index",
            "default": 0.5
          },
          "mbf": {
            "type": "number",
            "title": "Mbf",
            "description": "Mean time between failures",
            "default": 300.0
          },
          "ads": {
            "type": "number",
            "title": "Ads",
            "description": "Average Downtime Score",
            "default": 5.0
          },
          "ohi": {
            "type": "number",
            "title": "Ohi",
            "description": "Overall Health Index",
            "default": 80.0
          },
          "cmes": {
            "type": "number",
            "title": "Cmes",
            "description": "Cumulative Maintenance Effort Score",
            "default": 1000.0
          },
          "uer": {
            "type": "number",
            "title": "Uer",
            "description": "Utilisation Efficiency Ratio",
            "default": 0.5
          },
          "vehicle_type": {
            "type": "string",
            "title": "Vehicle Type",
            "description": "Vehicle subclass type",
            "default": "Light Truck"
          },
          "route_info": {
            "type": "string",
            "title": "Route Info",
            "description": "Primary route info",
            "default": "Urban Delivery"
          },
          "road_conditions": {
            "type": "string",
            "title": "Road Conditions",
            "description": "Typical road condition",
            "default": "Smooth"
          },
          "weather_conditions": {
            "type": "string",
            "title": "Weather Conditions",
            "description": "Typical weather condition",
            "default": "Clear"
          },
          "brake_condition": {
            "type": "string",
            "title": "Brake Condition",
            "description": "Brake pad health category",
            "default": "Good"
          }
        },
        "type": "object",
        "required": [
          "vehicle_age_years",
          "usage_hours",
          "fuel_consumption",
          "health_score"
        ],
        "title": "FleetPredictInput",
        "example": {
          "actual_load": 3000.0,
          "ads": 3.5,
          "anomalies_detected": 1,
          "brake_condition": "Good",
          "cmes": 1500.0,
          "days_since_last_maintenance": 45,
          "diagnostic_trouble_code_count": 0,
          "failure_history": 2,
          "fuel_consumption": 10.5,
          "fuel_per_hour": 1.25,
          "health_score": 80.0,
          "load_capacity": 5000.0,
          "load_utilization_pct": 60.0,
          "maintenance_cost": 1200.0,
          "mbf": 250.0,
          "ohi": 85.0,
          "pcr": 0.55,
          "predictive_score": 0.85,
          "road_conditions": "Smooth",
          "route_info": "Urban Delivery",
          "tpi": 0.7,
          "uer": 0.75,
          "uir": 0.6,
          "usage_hours": 8000.0,
          "vehicle_age_years": 8,
          "vehicle_type": "Light Truck",
          "weather_conditions": "Clear"
        }
      },
      "FleetPredictOutput": {
        "properties": {
          "ev_readiness_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Ev Readiness Score",
            "description": "Suitability score"
          },
          "readiness_category": {
            "type": "string",
            "title": "Readiness Category",
            "description": "Suitability category"
          },
          "model_used": {
            "type": "string",
            "title": "Model Used",
            "description": "Algorithm used for prediction"
          }
        },
        "type": "object",
        "required": [
          "ev_readiness_score",
          "readiness_category",
          "model_used"
        ],
        "title": "FleetPredictOutput"
      },
      "FleetSummaryOutput": {
        "properties": {
          "total_vehicles": {
            "type": "integer",
            "title": "Total Vehicles",
            "description": "Total active vehicles analyzed in fleet"
          },
          "high_readiness_count": {
            "type": "integer",
            "title": "High Readiness Count",
            "description": "Vehicles with score >= 0.6"
          },
          "medium_readiness_count": {
            "type": "integer",
            "title": "Medium Readiness Count",
            "description": "Vehicles with score between 0.4 and 0.6"
          },
          "low_readiness_count": {
            "type": "integer",
            "title": "Low Readiness Count",
            "description": "Vehicles with score < 0.4"
          },
          "readiness_percentage": {
            "type": "number",
            "title": "Readiness Percentage",
            "description": "Percentage of vehicles eligible for transition"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/FleetVehicleReadiness"
            },
            "type": "array",
            "title": "Recommendations",
            "description": "Sample high-readiness recommendations"
          }
        },
        "type": "object",
        "required": [
          "total_vehicles",
          "high_readiness_count",
          "medium_readiness_count",
          "low_readiness_count",
          "readiness_percentage",
          "recommendations"
        ],
        "title": "FleetSummaryOutput"
      },
      "FleetVehicleReadiness": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id",
            "description": "Vehicle unique identifier"
          },
          "vehicle_type": {
            "type": "string",
            "title": "Vehicle Type",
            "description": "Type/class of the vehicle"
          },
          "ev_readiness_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Ev Readiness Score",
            "description": "Electrification suitability index [0-1]"
          },
          "readiness_category": {
            "type": "string",
            "title": "Readiness Category",
            "description": "Categorized suitability status (e.g. High Readiness)"
          },
          "recommended_ev_replacement": {
            "type": "string",
            "title": "Recommended Ev Replacement",
            "description": "Suggested electric model replacement"
          },
          "estimated_cost_usd": {
            "type": "number",
            "title": "Estimated Cost Usd",
            "description": "Estimated vehicle acquisition cost in USD"
          },
          "lead_time_months": {
            "type": "integer",
            "title": "Lead Time Months",
            "description": "Delivery lead time in months"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "vehicle_type",
          "ev_readiness_score",
          "readiness_category",
          "recommended_ev_replacement",
          "estimated_cost_usd",
          "lead_time_months"
        ],
        "title": "FleetVehicleReadiness"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelRegistryOutput": {
        "properties": {
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelInfoOutput"
            },
            "type": "array",
            "title": "Models",
            "description": "List of all registered models"
          }
        },
        "type": "object",
        "required": [
          "models"
        ],
        "title": "ModelRegistryOutput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}